+++ tests/common.rs
@@ -76,10 +76,10 @@ impl TestData {
     }
 
     /// Creates a `TestData` instance from a test path, source file, and expected file.
-    /// 
-    /// 
+    ///
+    ///
     /// # Errors
-    /// 
+    ///
     /// Returns an error if the source or expected file cannot be read.
     pub fn from_path<P>(
         test_path: P,
@@ -146,11 +146,10 @@ pub(crate) fn test_readability(data: TestData) {
     check!("content", a_html, e_html, data.path);
 }
 
-
 /// Tests that the metadata is extracted correctly from the article.
-/// 
+///
 /// # Panics
-/// 
+///
 /// Panics if the test fails.
 pub fn test_metadata(data: TestData, host: Option<&str>) {
     let cfg = dom_smoothie::Config {
@@ -181,9 +180,9 @@ pub fn test_metadata(data: TestData, host: Option<&str>) {
 }
 
 /// Tests that the favicon is extracted correctly from the article metadata.
-/// 
+///
 /// # Panics
-/// 
+///
 /// Panics if the test fails.
 pub fn test_favicon<P>(test_path: P, host: Option<&str>, expected: Option<&str>)
 where