+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to the `dom_smoothie` crate will be documented in this file.
 ### Changed
 - Updated `dom_query`: `0.27.0` -> `0.28.0`.
 - Applied selected Clippy suggestions (pedantic) to improve code quality. 
+- Refactored relative URL resolution in the document.
 
 ## [0.17.0] - 2026-03-28